+++ router/cors.go
@@ -1,3 +1,6 @@
+// Copyright (C) 2025-2026 my-app-s
+// Licensed under the GNU AGPLv3
+
 package router
 
 import (